requirements.txt
python-digitalocean==1.15.0

